+++ src/morie/fn/qnorm.py
@@ -34,7 +34,7 @@ def qnorm(
         p_arr = np.exp(p_arr)
     if not lower_tail:
         p_arr = 1.0 - p_arr
-    return stats.norm(loc=mean, scale=sd).ppf(p_arr)
+    return np.scalar_out(p, stats.norm(loc=mean, scale=sd).ppf(p_arr))
 
 
 def cheatsheet() -> str: